requirements.txt
altgraph==0.17.4
asttokens==3.0.0
blinker==1.9.0
cachetools==4.2.4
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
decorator==5.1.1
exceptiongroup==1.2.2
executing==2.1.0
Flask==3.1.0
google-api-core==1.34.1
google-api-python-client==2.156.0
google-auth==2.37.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
google-cloud-core==2.4.1
google-cloud-translate==3.19.0
googleapis-common-protos==1.66.0
grpc-google-iam-v1==0.13.1
grpcio==1.68.1
grpcio-status==1.48.2
httplib2==0.22.0
idna==3.10
ipython==8.31.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.5
lxml==5.3.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
oauthlib==3.2.2
packaging==24.2
parso==0.8.4
pexpect==4.9.0
prompt_toolkit==3.0.48
proto-plus==1.25.0
protobuf==3.20.3
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
Pygments==2.18.0
pyinstaller==6.11.1
pyinstaller-hooks-contrib==2024.11
pyparsing==3.2.0
PyPDF2==3.0.1
python-docx==1.1.2
python-dotenv==1.0.1
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
six==1.17.0
stack-data==0.6.3
traitlets==5.14.3
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.3.0
wcwidth==0.2.13
Werkzeug==3.1.3
